HW05: Next remarks were implemented: - все проверки из AbstractArrayStorage переехали в наследники и дублируются, возможно стоит создать отд метод, который будет проверять индекс, и у каждого storage будет своя реализация такой проверки - MapStorage. deleteResumeByIndex/saveResumeByIndex/updateResumeByIndex проверяй индекс, не надо получать резюме из storage

+++ src/main/java/ru/javawebinar/basejava/storage/AbstractStorage.java
@@ -1,11 +1,14 @@
 package ru.javawebinar.basejava.storage;
 
+import ru.javawebinar.basejava.exception.ExistStorageException;
+import ru.javawebinar.basejava.exception.NotExistStorageException;
 import ru.javawebinar.basejava.model.Resume;
 
 import java.util.Objects;
 
 public abstract class AbstractStorage implements Storage {
 
+    protected abstract boolean isElementExistByIndex(Object indexObj);
     protected abstract void deleteResumeByIndex(Object indexObj);
     protected abstract Object getIndex(String uuid);
     protected abstract Resume getResumeByIndex(Object indexObj);
@@ -16,6 +19,9 @@ public abstract class AbstractStorage implements Storage {
     public void delete(String uuid) {
         Objects.requireNonNull(uuid, "ERROR delete. uuid cannot be null");
         Object indexObj = getIndex(uuid);
+        if (!isElementExistByIndex(indexObj)) {
+            throw new NotExistStorageException(uuid);
+        }
         deleteResumeByIndex(indexObj);
     }
 
@@ -23,6 +29,9 @@ public abstract class AbstractStorage implements Storage {
     public Resume get(String uuid) {
         Objects.requireNonNull(uuid, "ERROR get. uuid cannot be null");
         Object indexObj = getIndex(uuid);
+        if (!isElementExistByIndex(indexObj)) {
+            throw new NotExistStorageException(uuid);
+        }
         return getResumeByIndex(indexObj);
     }
 
@@ -30,6 +39,9 @@ public abstract class AbstractStorage implements Storage {
     public void save(Resume resume) {
         Objects.requireNonNull(resume, "ERROR save. Resume cannot be null");
         Object indexObj = getIndex(resume.getUuid());
+        if (isElementExistByIndex(indexObj)) {
+            throw new ExistStorageException(resume.getUuid());
+        }
         saveResumeByIndex(indexObj, resume);
     }
 
@@ -37,6 +49,9 @@ public abstract class AbstractStorage implements Storage {
     public void update(Resume resume) {
         Objects.requireNonNull(resume, "ERROR update. Resume cannot be null");
         Object indexObj = getIndex(resume.getUuid());
+        if (!isElementExistByIndex(indexObj)) {
+            throw new NotExistStorageException(resume.getUuid());
+        }
         updateResumeByIndex(indexObj, resume);
     }
